react-agent/pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
